requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
click==8.1.7
exceptiongroup==1.2.1
fastapi-slim==0.111.0
h11==0.14.0
hiredis==2.3.2
idna==3.7
pydantic==2.7.3
pydantic_core==2.18.4
redis==5.0.5
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.1
websockets==12.0
